LayoutTests/storage/indexeddb/removed-expected.txt

Ensure that some obsolete IndexedDB features are gone.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

PASS self.webkitIDBDatabaseError is undefined.
PASS 'IDBDatabaseException' in self is false
PASS 'errorCode' in indexedDB.open(dbname) is false
PASS 'setVersion' in IDBDatabase.prototype is false
PASS document.createEvent('IDBUpgradeNeededEvent') threw exception NotSupportedError: The operation is not supported..
PASS document.createEvent('IDBVersionChangeEvent') threw exception NotSupportedError: The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
